Renamed the services file that will handle all routes involving the user and their data.
Also applied fixes to all files that were using the old file.
