+++ src/tempProfile.jsx
@@ -2,7 +2,7 @@
 // This is a temporary profile component to test the auth0 authentication and authorization.
 
 import { useAuth0 } from '@auth0/auth0-react'
-import { fetchUserDetails } from './services/fetchUserDetails'
+import { fetchUserDetails } from './services/userService'
 
 export const Profile = () => {
   const { user, isAuthenticated, isLoading, getAccessTokenSilently } =
